+++ tests/Integration/WatcherClient.php
@@ -70,7 +70,7 @@ class WatcherClient extends AbstractClient
         $this->addDecision($now, '12h', '+12 hours', TestHelpers::BAD_IP, 'captcha');
         $this->addDecision($now, '24h', self::HOURS24, TestHelpers::BAD_IP . '/' . TestHelpers::IP_RANGE, 'ban');
         $this->addDecision($now, '24h', '+24 hours', TestHelpers::JAPAN, 'captcha', Constants::SCOPE_COUNTRY);
-        usleep(500 * 1000);
+        usleep(500 * 1000);      // 500ms
     }
 
     /** Set the second watcher state */